Downloads/SkillMatrix-Pro-main/ai-engine/requirements.txt
fastapi
uvicorn
PyPDF2
groq
motor
python-dotenv
pydantic
